Sample rewrite of road racing
This is a sample rewrite/reorganization of the road racing chapter as a
proof of concept for the separation of rider and host rules for clarity.
Most text was left unchanged, and only moved around to new places.  Some
small changes to the text were made, particularly in the more fundamental
rider rules.