Added rewrite of muni chapter
This commit covers a rewrite of the muni chapter, along the lines of what
was done for the road racing chapter.  Muni now takes up two chapters in
the rule book, which has caused some inconsistency with the file naming
conventions.  After the reorg is complete, the file names should be
normalized.